+++ src/ebic/crud/tomograms.py
@@ -1,15 +1,30 @@
+from typing import Literal
+
 from fastapi import HTTPException
+from sqlalchemy import Column
 from sqlalchemy import func as f
 
 from ..models.response import CtfTiltAlignList, FullMovieWithTilt, GenericPlot
 from ..models.table import CTF, MotionCorrection, Movie, TiltImageAlignment, Tomogram
 from ..utils.database import db, paginate
-from ..utils.generic import parse_json_file
-from .movies import get_tomogram_auto_proc_attachment
+from ..utils.generic import parse_json_file, validate_path
 
 
-def get_shift_plot(id: int):
-    data = parse_json_file(get_tomogram_auto_proc_attachment(id, "Graph"))
+def _get_generic_tomogram_file(tomogramId: int, column: Column) -> str:
+    return (
+        db.session.query(f.concat(Tomogram.fileDirectory, "/", column))
+        .filter(Tomogram.tomogramId == tomogramId)
+        .scalar()
+    )
+
+
+@validate_path
+def _get_shift_plot_path(tomogramId: int):
+    return _get_generic_tomogram_file(tomogramId, Tomogram.xyShiftPlot)
+
+
+def get_shift_plot(tomogramId: int):
+    data = parse_json_file(_get_shift_plot_path(tomogramId))
 
     if not data:
         raise HTTPException(status_code=404, detail="Invalid or empty data file")
@@ -61,3 +76,20 @@ def get_ctf(tomogramId: int):
     )
 
     return CtfTiltAlignList(items=data)
+
+
+@validate_path
+def get_slice_path(tomogramId: int) -> str:
+    return _get_generic_tomogram_file(tomogramId, Tomogram.centralSliceImage)
+
+
+@validate_path
+def get_movie_path(tomogramId: int) -> str:
+    return _get_generic_tomogram_file(tomogramId, Tomogram.tomogramMovie)
+
+
+@validate_path
+def get_projection_path(tomogramId: int, axis: Literal["xy", "xz"]) -> str:
+    return _get_generic_tomogram_file(
+        tomogramId, Tomogram.projXZ if axis == "xz" else Tomogram.projXY
+    )