+++ src/ebic/routes/tomograms.py
@@ -1,8 +1,10 @@
+from typing import Literal
+
 from fastapi import APIRouter, Depends
 from fastapi.responses import FileResponse
 
 from ..auth import Permissions
-from ..crud import movies, tomograms
+from ..crud import tomograms as crud
 from ..models.response import CtfTiltAlignList, FullMovieWithTilt, GenericPlot
 from ..utils.dependencies import pagination
 
@@ -17,7 +19,7 @@ router = APIRouter(
 @router.get("/{tomogramId}/shiftPlot", response_model=GenericPlot)
 def get_shift_plot(tomogramId: int = Depends(auth)):
     "Get X-Y shift plot data"
-    return tomograms.get_shift_plot(tomogramId)
+    return crud.get_shift_plot(tomogramId)
 
 
 @router.get("/{tomogramId}/motion", response_model=FullMovieWithTilt)
@@ -25,17 +27,29 @@ def get_motion_correction(
     tomogramId: int = Depends(auth), page: dict[str, int] = Depends(pagination)
 ):
     "Get motion correction data for the given tomogram"
-    return tomograms.get_motion_correction(tomogramId=tomogramId, **page)
+    return crud.get_motion_correction(tomogramId=tomogramId, **page)
 
 
 @router.get("/{tomogramId}/centralSlice", response_class=FileResponse)
 def get_slice(tomogramId: int = Depends(auth)):
     """Get tomogram central slice image"""
-    return movies.get_tomogram_auto_proc_attachment(tomogramId)
+    return crud.get_slice_path(tomogramId)
+
+
+@router.get("/{tomogramId}/movie", response_class=FileResponse)
+def get_movie(tomogramId: int = Depends(auth)):
+    """Get tomogram movie image"""
+    return crud.get_movie_path(tomogramId)
+
+
+@router.get("/{tomogramId}/projection", response_class=FileResponse)
+def get_projection(axis: Literal["xy", "xz"], tomogramId: int = Depends(auth)):
+    """Get tomogram projection image"""
+    return crud.get_projection_path(tomogramId=tomogramId, axis=axis)
 
 
 @router.get("/{tomogramId}/ctf", response_model=CtfTiltAlignList)
 def get_ctf(tomogramId: int = Depends(auth)):
     """Get astigmatism, resolution and defocus as a function of tilt image
     alignment refined tilt angles"""
-    return tomograms.get_ctf(tomogramId)
+    return crud.get_ctf(tomogramId)